+++ packages/pwa/src/components/devToolbar/RegisterFeedItemImporterDevTool.tsx
@@ -1,11 +1,11 @@
 import {useCallback, useEffect, useState} from 'react';
 
+import {PWA_FEED_SOURCE} from '@shared/lib/feedSources.shared';
 import {isValidUrl} from '@shared/lib/urls.shared';
 import {assertNever} from '@shared/lib/utils.shared';
 
 import {AsyncStatus} from '@shared/types/asyncState.types';
 import {DevToolbarSectionType} from '@shared/types/devToolbar.types';
-import {PWA_FEED_SOURCE} from '@shared/types/feedItems.types';
 
 import {useDevToolbarStore} from '@sharedClient/stores/DevToolbarStore';
